+++ fivetran/tests/mock/resource_connector_schema_allow_columns_test.go
@@ -135,7 +135,7 @@ func TestResourceSchemaDisableColumnMissingInSchemaResponseMock(t *testing.T) {
 
 		Check: resource.ComposeAggregateTestCheckFunc(
 			func(s *terraform.State) error {
-				assertEqual(t, getHandler.Interactions, 3)   // 1 read attempt before reload, 1 read after create
+				assertEqual(t, getHandler.Interactions, 2)   // 1 read attempt before reload, 1 read after create
 				assertEqual(t, patchHandler.Interactions, 1) // Update SCM and align schema
 				assertNotEmpty(t, schemaData)                // schema initialised
 				return nil
@@ -200,6 +200,589 @@ func TestResourceSchemaDisableColumnMissingInSchemaResponseMock(t *testing.T) {
 			ProtoV6ProviderFactories: ProtoV6ProviderFactories,
 			CheckDestroy: func(s *terraform.State) error {
 				// there is no possibility to destroy schema config - it alsways exists within the connector
+				return nil
+			},
+
+			Steps: []resource.TestStep{
+				step1,
+			},
+		},
+	)
+}
+
+func TestResourceSchemaEmptyColumnsListMock(t *testing.T) {
+
+	var schemaData map[string]interface{}
+
+	var getHandler *mock.Handler
+	var patchHandler *mock.Handler
+
+	var columnsListResponse = `
+{
+	"columns": {
+		"column_1": {
+			"name_in_destination": "column_1",
+			"enabled": true,
+			"hashed": false,
+			"enabled_patch_settings": {
+				"allowed": true
+			}
+		},
+		"column_2": {
+			"name_in_destination": "column_2",
+			"enabled": true,
+			"hashed": false,
+			"enabled_patch_settings": {
+				"allowed": true
+			}
+		}
+	}
+}`
+	var schemasWoColumnsJsonResponse = `
+{
+	"enable_new_by_default": true,
+	"schemas": {
+		"schema_1": {
+			"name_in_destination": "schema_1",
+			"enabled": true,
+			"tables": {
+				"table_1": {
+					"name_in_destination": "table_1",
+					"sync_mode": "LIVE",
+					"enabled": true,
+					"enabled_patch_settings": {
+						"allowed": true
+					},
+					"columns": {
+						"column_1": {
+							"name_in_destination": "column_1",
+							"enabled": true,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						},
+						"column_2": {
+							"name_in_destination": "column_2",
+							"enabled": true,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						},
+						"column_3": {
+							"name_in_destination": "column_3",
+							"enabled": true,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						}
+					}
+				},
+				"table_2": {
+					"name_in_destination": "table_2",
+					"sync_mode": "LIVE",
+					"enabled": true,
+					"enabled_patch_settings": {
+						"allowed": true
+					},
+					"columns": {
+						"column_2_1": {
+							"name_in_destination": "column_2_1",
+							"enabled": false,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						},
+						"column_2_2": {
+							"name_in_destination": "column_2_2",
+							"enabled": true,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						}
+					}
+				}
+			}
+		}
+	},
+	"schema_change_handling": "%v"
+}`
+	var schemaWithColumnJsonResponse = `
+{
+	"schema_change_handling": "ALLOW_COLUMNS",
+	"schemas": {
+		"schema_1": {
+			"name_in_destination": "schema_1",
+			"enabled": true,
+			"tables": {
+				"table_1": {
+					"name_in_destination": "table_1",
+					"enabled": true,
+					"sync_mode": "SOFT_DELETE",
+					"enabled_patch_settings": {
+						"allowed": true
+					},
+					"columns": {
+						"column_1": {
+							"name_in_destination": "column_1",
+							"enabled": false,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						},
+						"column_2": {
+							"name_in_destination": "column_2",
+							"enabled": false,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						},
+						"column_3": {
+							"name_in_destination": "column_3",
+							"enabled": true,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						}
+					}
+				},
+				"table_2": {
+					"name_in_destination": "table_2",
+					"enabled": true,
+					"sync_mode": "SOFT_DELETE",
+					"enabled_patch_settings": {
+						"allowed": true
+					},
+					"columns": {
+						"column_2_1": {
+							"name_in_destination": "column_2_1",
+							"enabled": false,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						},
+						"column_2_2": {
+							"name_in_destination": "column_2_2",
+							"enabled": true,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						}
+					}
+				}
+			}
+		}
+	}
+}`
+
+	step1 := resource.TestStep{
+		Config: `
+			locals {
+				tables = {
+					"table_1" = {
+						name = "table_1"
+						disabled_columns      = ["column_1", "column_2"]
+					}
+					"table_2" = {
+						name = "table_2"
+						disabled_columns      = [] # empty list
+					}
+				}
+			}
+
+			resource "fivetran_connector_schema_config" "test_schema" {
+				provider = fivetran-provider
+
+				connector_id           = "connector_id"
+				schema_change_handling = "ALLOW_COLUMNS"
+
+				schemas = {
+					"schema_1" = {
+						enabled = true
+						tables = {
+							for table in local.tables : table.name => {
+								enabled   = true
+								sync_mode = "SOFT_DELETE"
+								columns = {
+									for column in table.disabled_columns : column => {
+										enabled = false
+									}
+								}
+							}
+						}
+					}
+				}
+			}`,
+
+		Check: resource.ComposeAggregateTestCheckFunc(
+			func(s *terraform.State) error {
+				assertEqual(t, getHandler.Interactions, 2)   // 1 read attempt before reload, 1 read after create
+				assertEqual(t, patchHandler.Interactions, 1) // Update SCM and align schema
+				assertNotEmpty(t, schemaData)                // schema initialised
+				return nil
+			},
+			resource.TestCheckResourceAttr("fivetran_connector_schema_config.test_schema", "schema_change_handling", "ALLOW_COLUMNS"),
+			resource.TestCheckResourceAttr("fivetran_connector_schema_config.test_schema", "schemas.schema_1.tables.table_1.enabled", "true"),
+			resource.TestCheckResourceAttr("fivetran_connector_schema_config.test_schema", "schemas.schema_1.tables.table_1.sync_mode", "SOFT_DELETE"),
+			resource.TestCheckResourceAttr("fivetran_connector_schema_config.test_schema", "schemas.schema_1.tables.table_2.enabled", "true"),
+			resource.TestCheckResourceAttr("fivetran_connector_schema_config.test_schema", "schemas.schema_1.tables.table_2.sync_mode", "SOFT_DELETE"),
+
+			resource.TestCheckResourceAttr("fivetran_connector_schema_config.test_schema", "schemas.schema_1.tables.table_1.columns.%", "2"),
+			resource.TestCheckResourceAttr("fivetran_connector_schema_config.test_schema", "schemas.schema_1.tables.table_1.columns.column_1.enabled", "false"),
+			resource.TestCheckResourceAttr("fivetran_connector_schema_config.test_schema", "schemas.schema_1.tables.table_1.columns.column_2.enabled", "false"),
+
+			resource.TestCheckResourceAttr("fivetran_connector_schema_config.test_schema", "schemas.schema_1.tables.table_2.columns.%", "0"),
+		),
+	}
+
+	resource.Test(
+		t,
+		resource.TestCase{
+			PreCheck: func() {
+				mockClient.Reset()
+				schemaData = nil
+
+				getHandler = mockClient.When(http.MethodGet, "/v1/connections/connector_id/schemas").ThenCall(
+					func(req *http.Request) (*http.Response, error) {
+						if nil == schemaData {
+							schemaData = createMapFromJsonString(t, fmt.Sprintf(schemasWoColumnsJsonResponse, "ALLOW_ALL"))
+						}
+						return fivetranSuccessResponse(t, req, http.StatusOK, "Success", schemaData), nil
+					},
+				)
+
+				mockClient.When(http.MethodGet, "/v1/connections/connector_id/schemas/schema_1/tables/table_1/columns").ThenCall(
+					func(req *http.Request) (*http.Response, error) {
+						return fivetranSuccessResponse(t, req, http.StatusOK, "Success", createMapFromJsonString(t, columnsListResponse)), nil
+					},
+				)
+
+				patchHandler = mockClient.When(http.MethodPatch, "/v1/connections/connector_id/schemas").ThenCall(
+					func(req *http.Request) (*http.Response, error) {
+						body := 
+						requestBodyToJson(t, req)
+
+						assertEqual(t, body["schema_change_handling"], "ALLOW_COLUMNS")
+
+						schemas := assertKeyExists(t, body, "schemas").(map[string]interface{})
+
+						schema := assertKeyExists(t, schemas, "schema_1").(map[string]interface{})
+
+						AssertKeyDoesNotExist(t, schema, "enabled")
+						tables := assertKeyExists(t, schema, "tables").(map[string]interface{})
+
+						assertKeyExists(t, tables, "table_1")
+						assertKeyExists(t, tables, "table_2")
+
+						// create schema structure
+						schemaData = createMapFromJsonString(t, schemaWithColumnJsonResponse)
+
+						return fivetranSuccessResponse(t, req, http.StatusOK, "Success", schemaData), nil
+					},
+				)
+			},
+			ProtoV6ProviderFactories: ProtoV6ProviderFactories,
+			CheckDestroy: func(s *terraform.State) error {
+				// there is no possibility to destroy schema config - it always exists within the connector
+				return nil
+			},
+
+			Steps: []resource.TestStep{
+				step1,
+			},
+		},
+	)
+}
+
+func TestResourceSchemaEmptyColumnsListOldSchemaMock(t *testing.T) {
+
+	var schemaData map[string]interface{}
+
+	var getHandler *mock.Handler
+	var patchHandler *mock.Handler
+
+	var columnsListResponse = `
+{
+	"columns": {
+		"column_1": {
+			"name_in_destination": "column_1",
+			"enabled": true,
+			"hashed": false,
+			"enabled_patch_settings": {
+				"allowed": true
+			}
+		},
+		"column_2": {
+			"name_in_destination": "column_2",
+			"enabled": true,
+			"hashed": false,
+			"enabled_patch_settings": {
+				"allowed": true
+			}
+		}
+	}
+}`
+	var schemasWoColumnsJsonResponse = `
+{
+	"enable_new_by_default": true,
+	"schemas": {
+		"schema_1": {
+			"name_in_destination": "schema_1",
+			"enabled": true,
+			"tables": {
+				"table_1": {
+					"name_in_destination": "table_1",
+					"sync_mode": "LIVE",
+					"enabled": true,
+					"enabled_patch_settings": {
+						"allowed": true
+					},
+					"columns": {
+						"column_1": {
+							"name_in_destination": "column_1",
+							"enabled": true,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						},
+						"column_2": {
+							"name_in_destination": "column_2",
+							"enabled": true,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						},
+						"column_3": {
+							"name_in_destination": "column_3",
+							"enabled": true,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						}
+					}
+				},
+				"table_2": {
+					"name_in_destination": "table_2",
+					"sync_mode": "LIVE",
+					"enabled": true,
+					"enabled_patch_settings": {
+						"allowed": true
+					},
+					"columns": {
+						"column_2_1": {
+							"name_in_destination": "column_2_1",
+							"enabled": false,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						},
+						"column_2_2": {
+							"name_in_destination": "column_2_2",
+							"enabled": true,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						}
+					}
+				}
+			}
+		}
+	},
+	"schema_change_handling": "%v"
+}`
+	var schemaWithColumnJsonResponse = `
+{
+	"schema_change_handling": "ALLOW_COLUMNS",
+	"schemas": {
+		"schema_1": {
+			"name_in_destination": "schema_1",
+			"enabled": true,
+			"tables": {
+				"table_1": {
+					"name_in_destination": "table_1",
+					"enabled": true,
+					"sync_mode": "SOFT_DELETE",
+					"enabled_patch_settings": {
+						"allowed": true
+					},
+					"columns": {
+						"column_1": {
+							"name_in_destination": "column_1",
+							"enabled": false,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						},
+						"column_2": {
+							"name_in_destination": "column_2",
+							"enabled": false,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						},
+						"column_3": {
+							"name_in_destination": "column_3",
+							"enabled": true,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						}
+					}
+				},
+				"table_2": {
+					"name_in_destination": "table_2",
+					"enabled": true,
+					"sync_mode": "SOFT_DELETE",
+					"enabled_patch_settings": {
+						"allowed": true
+					},
+					"columns": {
+						"column_2_1": {
+							"name_in_destination": "column_2_1",
+							"enabled": false,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						},
+						"column_2_2": {
+							"name_in_destination": "column_2_2",
+							"enabled": true,
+							"hashed": false,
+							"enabled_patch_settings": {
+								"allowed": true
+							}
+						}
+					}
+				}
+			}
+		}
+	}
+}`
+
+	step1 := resource.TestStep{
+		Config: `
+			locals {
+				tables = {
+					"table_1" = {
+						name = "table_1"
+						disabled_columns      = ["column_1", "column_2"]
+					}
+					"table_2" = {
+						name = "table_2"
+						disabled_columns      = [] # empty list
+					}
+				}
+			}
+
+			resource "fivetran_connector_schema_config" "test_schema" {
+				provider = fivetran-provider
+
+				connector_id           = "connector_id"
+				schema_change_handling = "ALLOW_COLUMNS"
+
+				schema {
+					name = "schema_1"
+					dynamic "table" {
+						for_each = local.tables
+						iterator = tables
+						content {
+							name      = tables.key
+							sync_mode = "SOFT_DELETE"
+							enabled   = true
+							dynamic "column" {
+								for_each = tables.value.disabled_columns
+								content {
+									name    = column.value
+									enabled = false
+								}
+							}
+						}
+					}
+				}
+			}`,
+
+		Check: resource.ComposeAggregateTestCheckFunc(
+			func(s *terraform.State) error {
+				assertEqual(t, getHandler.Interactions, 2)   // 1 read attempt before reload, 1 read after create
+				assertEqual(t, patchHandler.Interactions, 1) // Update SCM and align schema
+				assertNotEmpty(t, schemaData)                // schema initialised
+				return nil
+			},
+			resource.TestCheckResourceAttr("fivetran_connector_schema_config.test_schema", "schema_change_handling", "ALLOW_COLUMNS"),
+			resource.TestCheckResourceAttr("fivetran_connector_schema_config.test_schema", "schema.0.table.0.enabled", "true"),
+			resource.TestCheckResourceAttr("fivetran_connector_schema_config.test_schema", "schema.0.table.0.sync_mode", "SOFT_DELETE"),
+			resource.TestCheckResourceAttr("fivetran_connector_schema_config.test_schema", "schema.0.table.1.enabled", "true"),
+			resource.TestCheckResourceAttr("fivetran_connector_schema_config.test_schema", "schema.0.table.1.sync_mode", "SOFT_DELETE"),
+		),
+	}
+
+	resource.Test(
+		t,
+		resource.TestCase{
+			PreCheck: func() {
+				mockClient.Reset()
+				schemaData = nil
+
+				getHandler = mockClient.When(http.MethodGet, "/v1/connections/connector_id/schemas").ThenCall(
+					func(req *http.Request) (*http.Response, error) {
+						if nil == schemaData {
+							schemaData = createMapFromJsonString(t, fmt.Sprintf(schemasWoColumnsJsonResponse, "ALLOW_ALL"))
+						}
+						return fivetranSuccessResponse(t, req, http.StatusOK, "Success", schemaData), nil
+					},
+				)
+
+				mockClient.When(http.MethodGet, "/v1/connections/connector_id/schemas/schema_1/tables/table_1/columns").ThenCall(
+					func(req *http.Request) (*http.Response, error) {
+						return fivetranSuccessResponse(t, req, http.StatusOK, "Success", createMapFromJsonString(t, columnsListResponse)), nil
+					},
+				)
+
+				patchHandler = mockClient.When(http.MethodPatch, "/v1/connections/connector_id/schemas").ThenCall(
+					func(req *http.Request) (*http.Response, error) {
+						body := 
+						requestBodyToJson(t, req)
+
+						assertEqual(t, body["schema_change_handling"], "ALLOW_COLUMNS")
+
+						schemas := assertKeyExists(t, body, "schemas").(map[string]interface{})
+
+						schema := assertKeyExists(t, schemas, "schema_1").(map[string]interface{})
+
+						AssertKeyDoesNotExist(t, schema, "enabled")
+						tables := assertKeyExists(t, schema, "tables").(map[string]interface{})
+
+						assertKeyExists(t, tables, "table_1")
+						assertKeyExists(t, tables, "table_2")
+
+						// create schema structure
+						schemaData = createMapFromJsonString(t, schemaWithColumnJsonResponse)
+
+						return fivetranSuccessResponse(t, req, http.StatusOK, "Success", schemaData), nil
+					},
+				)
+			},
+			ProtoV6ProviderFactories: ProtoV6ProviderFactories,
+			CheckDestroy: func(s *terraform.State) error {
+				// there is no possibility to destroy schema config - it always exists within the connector
 				return nil
 			},